pyproject.toml
[tool.poetry]
name = "kiroku"
version = "0.1.0"
description = ""
authors = ["Ryosuke Imai <imai@qunasys.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
mypy = "^0.910"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
